pequenas alterações
pequenas configuraçoes no formulário
