+++ src/components/Steps.css
@@ -6,7 +6,8 @@
     max-width: 400px;
     margin: 0 auto;
     position: relative;
-    margin-bottom: 2rem;
+    margin-bottom: 1rem;
+    padding: auto;
 }
 
 .steps::after {
@@ -22,7 +23,7 @@
     background-color: #000000;
     z-index: 1;
     width: 120px;
-    padding: .5rem;
+    padding: 0.5rem;
 }
 
 .step svg{
@@ -31,7 +32,7 @@
 }
 
 .active {
-font-weight: bold;
+    font-weight: bold;
 }
 
 .active > svg {
