+++ src/components/Thanks.jsx
@@ -7,16 +7,14 @@ const Thanks = () => {
             <label htmlFor="DataHora">Data e Horário </label>
             <input type="datetime-local" name="daytime" id='daytime' required/>
         </div>
-        <div className="form-control">
-            <p></p>
-            <label htmlFor="legenda">
-            <input type="checkbox" name="" id='daytime' required/>
-            Ative para receber promoções e informações do nosso serviço
+        <div className="checkbox">
+            <input type="checkbox" name="checkbox" id='daytime' required/> <label>
+            <span class="check_new_btn">Ative para receber promoção e informação do nosso serviço.</span>
             </label>
         </div>
-        <div className="form-control">
-            <label htmlFor="mensagem">Comentário:</label><br />
-            <textarea name='mensagem' placeholder='Deixa seu feedback! 😉'></textarea>
+        <div className="form-control"><br/>
+            <label htmlFor="mensagem">Comentário:</label>
+            <textarea name='mensagem' placeholder='Digite aqui sua sua mensagem! 😉'></textarea>
         </div>
     </div>
     );
