+++ src/routes/TRANSformacao.jsx
@@ -1,54 +1,56 @@
-import React from 'react'
+//Components
+import {GrFormNext, GrFormPrevious} from 'react-icons/gr';
+import {FiSend} from 'react-icons/fi';
+import UserForm from "../components/UserForm";
+import ReviewForm from '../components/ReviewForm';
+import Thanks from '../components/Thanks';
+import Steps from '../components/Steps';
 
-const TRANSformacao = () => {
-    return (
-    <div>
-        <h2>TRANSformação</h2>
-        <div className="form-control">
-            <label htmlFor="name">Nome e Sobrenome: </label>
-            <input type="text" name="nomeSobre" id='nomeSobre' maxLength="20" placeholder='Digite aqui' required/>
-        </div>
+//Hooks
+import { useForm } from '../hooks/useForm';
+import { useState } from 'react';
 
-        <div className="form-control">
-            <label htmlFor="name">Nome Social: </label>
-            <input type="text" name="nameSocial" id='nameSocial' placeholder='Digite aqui' required/>
-        </div>
 
-        <div className="form-control">
-            <label htmlFor="sexo">Gênero: </label>
-            <select name="assunto" id="opcoes">
-                <option disabled="dissabled" selected="selected">-- Selecione a opção</option>
-                <option>Mulher (Trans ou Cis)</option>
-                <option>Homer (Trans ou Cis)</option>
-                <option>Gênero Não-Binário</option>
-            </select>
-        </div>
+import './Formulario.css';
 
-        <div className="form-control">
-            <label htmlFor="name">CPF: </label>
-            <input type="text" name="cpf" id='campo_cpf' size='11'maxLength='11' placeholder='Digite seu CPF aqui' required/>
-        </div>
+function App() {
+  const formComponents = [<UserForm/>, <ReviewForm/>, <Thanks/>];
 
-        <div className="form-control">
-            <label htmlFor="name">Idade: </label>
-            <input type="number" name="quantidade" id='idade' min="15" max="99" required/>
-        </div>
+  const {currentStep, currentComponent, changeStep, isLastStep, isFirstStep } = useForm(formComponents);
 
-        <div className="form-control">
-            <label htmlFor="email">E-mail: </label>
-            <input type="email" name="email" id='email' size="40" placeholder='seuemail@email.com' required/>
-        </div>
-
-        <div className="form-control">
-        <label htmlFor="contato">Contato: </label>
-            <label htmlFor="codigo">DDD </label>
-            <input type="text" name="codigo_area" id='DDD' size="2" maxLength="2" placeholder='xx' required/> 
-            <label htmlFor="celular">Celular </label>
-            <input type="text" name="number" id='celular' size="9" maxLength="9" placeholder='xxxxx-xxxx' required/>
-        </div>
-        
+  return (
+    <div className="app">
+      <div className="header">
+        <h2><strong>Realce o que já <i>é belo em Você</i></strong></h2>
+      </div>
+      <div className="form-container">
+        <p>Cadastra-se e Agenda!</p>
+        <Steps currentStep = { currentStep } />
+        <form onSubmit={(e) => changeStep(currentStep + 1, e)}>
+          <div className="inputs-container">{currentComponent}</div>
+          <div className="actions">
+            {!isFirstStep && (
+              <button type="button" onClick={() => changeStep(currentStep - 1)}>
+              <GrFormPrevious/>
+              <span>Volta</span>
+            </button>
+            )}
+            {!isLastStep ? (
+              <button type="submit">
+              <span>Avançar</span>
+              <GrFormNext/>
+            </button>
+            ) : (
+              <button type="button">
+              <FiSend/>
+              <span>Enviar</span>
+            </button>
+            )}
+          </div>
+        </form>
+      </div>
     </div>
-  )
+  );
 }
 
-export default TRANSformacao
+export default App;
